ListTracks refactored to modular and changed to Playlist; deleted temp: favorite page, favorite btn, your uploads page; player modular refactoring started
