+++ src/modules/Player/components/PlayerTrackInfo.vue
@@ -1,6 +1,6 @@
 <script setup>
 import { useIsDesktop } from "@/shared/helpers/useIsDesktop.js";
-import { usePlayerStore } from "@/stores/PlayerStore.js";
+import { usePlayerStore } from "@/modules/Player/store/PlayerStore.js";
 import { storeToRefs } from "pinia";
