+++ src/modules/Playlists/components/PlaylistSelector.vue
@@ -1,0 +1,11 @@
+<script setup>
+
+</script>
+
+<template>
+
+</template>
+
+<style scoped>
+
+</style>
