+++ src/modules/Playlists/index.js
@@ -1,0 +1,2 @@
+export { default as PlaylistSelector } from './components/PlaylistSelector.vue';
+export { default as Playlist } from './components/Playlist.vue';
