+++ src/pages/Register.vue
@@ -1,12 +1,7 @@
 <script setup>
-import RegisterForm from '../modules/users/RegisterForm.vue';
+import RegisterForm from '../modules/RegisterForm/RegisterForm.vue';
 </script>
 
 <template>
-    <main class="col-12 p-5 mb-5 min-vh-100 d-flex justify-content-center align-items-center ">
-        <div class="bg-1E p-4 rounded shadow ">
-            <h2 class="color-E0">Зарегистрируйте аккаунт</h2>
-            <RegisterForm/>
-        </div>
-    </main>
+  <RegisterForm/>
 </template>
