+++ src/ui/LikeButton.vue
@@ -1,6 +1,5 @@
 <script setup>
 
-
 </script>
 
 <template>
